refactor MMKV backup data to util
