+++ src/services/backup/drive/restoreTasks.ts
@@ -3,7 +3,7 @@ import { BackupDataFileName, RestoreTask, TaskType } from '../types';
 import { BackupCategory, BackupNovel } from '@database/types';
 import { _restoreNovelAndChapters } from '@database/queries/NovelQueries';
 import { _restoreCategory } from '@database/queries/CategoryQueries';
-import { restoreMMKVData } from '../remote/helpers';
+import { restoreMMKVData } from '../utils';
 
 export const restoreNovel = (
   parentId: string,
